refactor(be): modules structure
