+++ backend-nest/src/modules/prisma/prisma.module.ts
@@ -1,0 +1,12 @@
+import { Global, Module } from "@nestjs/common";
+
+import { PrismaService } from "src/modules/prisma/prisma.service";
+
+@Global()
+@Module({
+    imports: [],
+    controllers: [],
+    providers: [PrismaService],
+    exports: [PrismaService],
+})
+export class PrismaModule {}
